==> scripts/data_wrangling/1a_load_editors.ipynb <==
import pandas as pd
import numpy as np
import wmfdata as wmf
from wmfdata import charting, mariadb, hive, spark
from wmfdata.utils import pct_str, pd_display_all

#clear cache, restart kernel, then clear cache again then run the pip install below
rm -rf ~/.cache/pip/http/*

pip install --upgrade git+https://github.com/neilpquinn/wmfdata.git@hive-engines

## Read data

### Get data 2018

#2018 
t1_recipients =  pd.read_csv("../../data/raw/editors/2018/2018_grant_recipients.csv", sep=',', encoding = 'utf-8')

#ptp+ jury names
idc_jury_clean = pd.read_csv("../../data/processed/contest_api/indic_contests_jury_data.csv", sep=',', encoding = 'utf-8')

#ptp+ editor names
idc = pd.read_csv("../../data/processed/contest_api/indic_contests_data_clean.csv", sep=',', encoding = 'utf-8')

### Get data 2019

participants = pd.read_csv("../../data/processed/participants_list_received_clean/2019/participants.csv", sep=',', encoding = 'utf-8')
#editors_report_articles_per_month
erapm = pd.read_csv("../../data/raw/articles/2019/editors_report_articles_per_month.csv", sep=',', encoding = 'utf-8')

import glob
import pandas as pd

files = glob.glob('../../data/raw/editors/2019/Indonesia/*.csv')
dfs = []
for file in files:
    df = pd.read_csv(file)
    df['filename'] = file
    dfs.append(df)
editors_raw = pd.concat(dfs, ignore_index=True)

editors = editors_raw[['User name', 'Wiki', 'filename','Gender', 'Received incentive', 'Type of incentive']]

editors = editors.rename(columns={'Wiki':'language_name',
                                    'User name': 'user_name',
                                    'Received incentive'  : 'received_incentive',
                                    'Type of incentive':'incentive_type',
                       })

#replace strings to comply with db table values
editors = editors.replace({'language_name' : { 'Bahasa Indonesia' : 'Indonesian',
                                                 'Sunda'            : 'Sundanese',
                                                 'Jawa'             : 'Javanese'
                                               }
                            })


#clean up typos in the the incentive type column
editors['incentive_type'] = editors['incentive_type'].str.replace('Laptop and internet' , 'laptop_internet')
editors['incentive_type'] = editors['incentive_type'].str.replace('Laptop dan internet' , 'laptop_internet')
editors['incentive_type'] = editors['incentive_type'].str.lower()

#create a column for GLOW participation
editors['glow_editor_signup'] = True

editors['received_incentive'] = editors['received_incentive'].replace(to_replace=['No', 'Yes'], value=[0, 1])

#handle column for 'Received incentive?'
editors['received_incentive'] = editors['received_incentive'].astype('bool')


editors['user_name'] = editors['user_name'].replace(' ', '_', regex=True)

lang_names =tuple(editors['language_name'].unique())

ci = wmf.hive.run("""
SELECT  language_code, database_code, language_name
FROM canonical_data.wikis
WHERE language_name IN {lang_names} AND database_group = 'wikipedia'
""".format(lang_names=lang_names))

#merge
df = editors.merge(ci, how="left", on=['language_name'])

df.to_csv("../../data/processed/participants_list_received_clean/indonesia/Saraswati1.csv", sep=',', encoding = 'utf-8', index=False)

## Clean data

#idc_s = idc.loc[idc.str.startswith('project-tiger-2018-', na=False)]
idc_ptp18 = idc[idc['contest_code'].str.contains('project-tiger-2018-')]
idc_ptp18_editors = list(idc_ptp18['user_name'].unique())

len(idc_ptp18_editors)

idc_ptp18b = idc_ptp18.copy()
idc_ptp18b = idc_ptp18b[['article_name', 'user_name']]

idc_ptp18b = idc_ptp18b.rename(columns={'user_name': 'username',
                                        'article_name': 'page_title',
                                        })

idc_ptp18_editors = idc_ptp18b[['username', 'glow_editor_signup']]

erapm2018  = idc_ptp18b[['username', 'page_title']].groupby('username').count().sort_values(by='page_title', ascending=False).rename(columns={'page_title':'article_count'}).reset_index(inplace=True)
erapm2018['glow_editor_signup'] = True

len(erapm2018)

#create bool column for glow_editor_signup using Fountain data in idc_ptp18_editors list
t1_recipients["glow_editor_signup"]=np.where(t1_recipients["username"].isin(idc_ptp18_editors), 1,0)
t1_recipients['glow_editor_signup'] = t1_recipients['glow_editor_signup'].astype('bool')

#create column for recieved incentive
t1_recipients['incentive_recipient'] = True

t1_recipients.duplicated(subset='username').sum()

t1_recipients_clean = t1_recipients[['username', 'incentive_recipient', 'incentive_type', 'glow_editor_signup']]

idc_ptp18[idc_ptp18['user_name']=='zo']

t1_recipients_clean.to_csv("../../data/processed/participants_list_received_clean/2018/t1_recipients_clean.csv", sep=',', encoding = 'utf-8', index=False)

erapm2018.to_csv("../../data/processed/participants_list_received_clean/2018/erapm2018.csv", sep=',', encoding = 'utf-8', index=False)

### Method 1

merge both dfs
ptp18_participants = pd.merge(t1_recipients_clean, erapm2018, how='outer', on='username')

### Method 2

#due to various script (Indic language) specific particulars it is easier to merge  and check rows by hand offline

participants_tally_2018 = pd.read_csv("../../data/processed/participants_list_received_clean/2018/PTP_participants_grantees_article_counts.csv", sep=',', encoding = 'utf-8')

cols1 = ['incentive_recipient', 'glow_editor_signup']
cols2 = ['incentive_type', 'article_count']

participants_tally_2018[cols1] = participants_tally_2018[cols1].fillna(False)
participants_tally_2018[cols2] = participants_tally_2018[cols2].fillna(0)

participants_tally_2018_clean = participants_tally_2018.copy()

participants_tally_2018_clean.to_csv("../../data/processed/participants_list_received_clean/2018/PTP_participants_grantees_article_counts_clean.csv", sep=',', encoding = 'utf-8', index=False)

## Clean data

### Method 1

tp_filepath = "../../data/raw/editors/2019/tiger2_participants_report/total_participants.csv"
sr_filepath = "../../data/raw/editors/2019/tiger2_participants_report/support_recipients.csv"

#participants.to_csv(filepath, index=False, header=False)
contributors = pd.read_csv(tp_filepath)
support_recipients = pd.read_csv(sr_filepath)

#TO DO - make all names lower case and care for double entries where incentive is diff

#rename column for later merging
contributors = contributors.rename(columns={'Which incentive?':'incentive_type',
                                            'Recieved incentive?': 'incentive_recipient',
                                            'Username': 'username'
                                            })

support_recipients = support_recipients.rename(columns={'Recieved incentive?': 'received_incentive',
                                                        'Username': 'username',
                                                        'Incentive type':'incentive_type'
                                                       })

#clean up typos in the the incentive type column
contributors['incentive_type'] = contributors['incentive_type'].str.replace('Laptop, Internet' , 'laptop_internet')
contributors['incentive_type'] = contributors['incentive_type'].str.replace('Laptop,Internet' , 'laptop_internet')
support_recipients['incentive_type'] = support_recipients['incentive_type'].str.replace('Laptop,Internet' , 'laptop_internet')

#create a column for GLOW participation
contributors['glow_editor_signup'] = True

contributors['received_incentive'] = contributors['received_incentive'].replace(to_replace=['No', 'Yes'], value=[0, 1])

#handle column for 'Received incentive?'
contributors['received_incentive'] = contributors['received_incentive'].astype('bool')
support_recipients['received_incentive'] = support_recipients['received_incentive'].astype('bool')

contributors['incentive_type'].unique()

support_recipients['incentive_type'].unique()

#identify editors not in editor list
e_p = list(contributors['username'])
e_r = list(support_recipients['username'])

support_recipients_missing = list(set(e_r)-set(e_p))
srta_r = support_recipients.loc[support_recipients['username'].isin(support_recipients_missing)]
srta = srta_r.copy()

#srta['glow_editor_signup'] = False

participants = pd.concat([contributors, srta], sort=True).fillna(False)

### Method 2

replacements = {
    'incentive_type' : {
        'Laptop,Internet' : 'laptop_internet',
        'Laptop, Internet' : 'laptop_internet',
    }
}

participants.replace(replacements, inplace=True)

#lowercase values
participants['incentive_type'] = participants['incentive_type'].str.lower()

#handle column for 'Received incentive?'
participants = participants.rename(columns={'received_incentive': 'incentive_recipient'})


participants['incentive_recipient'] = participants['incentive_recipient'].replace(to_replace=['No', 'Yes'], value=[0, 1])
participants['incentive_recipient'] = participants['incentive_recipient'].astype('bool')

editor_tallies = erapm [['user', 'Grand Total']]

editor_tallies = editor_tallies.rename(columns={'user': 'username',
                                                'Grand Total': 'total_articles'
                                                       })

participants_tally_2019 = pd.merge(participants, editor_tallies, how='left', on='username').fillna(0)

participants_tally_2019.to_csv("../../data/processed/participants_list_received_clean/2019/participants_tally_2019.csv", sep=',', encoding = 'utf-8', index=False)

## Preparing to create table and load into hive

participants2018 = participants_tally_2018[['glow_editor_signup', 'incentive_recipient', 'incentive_type', 'username']]
participants2019 = participants_tally_2019[['glow_editor_signup', 'incentive_recipient', 'incentive_type', 'username']]

participants2018.to_csv("../../data/processed/participants_list_received_clean/final/participants2018.csv", sep=',', encoding = 'utf-8', index=False)

participants2019.to_csv("../../data/processed/participants_list_received_clean/final/participants2019.csv", sep=',', encoding = 'utf-8', index=False)

!hive -e "Drop TABLE IF EXISTS florez.glow_tiger_participants"

!hive -e LOAD DATA LOCAL INPATH "../../data/processed/participants_list_received_clean/2019/participants_recipients_all.csv" 
INTO TABLE florez.glow_tiger_participants;

# Getting a list of participants into Hive

participants2019 = "../../data/processed/participants_list_received_clean/final/participants2019.csv"
hive.load_csv(
    participants2019,
    field_spec="glow_editor_signup boolean, incentive_recipient boolean, incentive_type string, username string",
    db_name="florez",
    table_name="glow_tiger_participants",
    sep=','
)

participants2018 = "../../data/processed/participants_list_received_clean/final/participants2018.csv"
hive.load_csv(
    participants2018,
    field_spec="glow_editor_signup boolean, incentive_recipient boolean, incentive_type string, username string",
    db_name="florez",
    table_name="glow_tiger_pilot_participants",
    sep=','
)
