==> scripts/data_wrangling/1b_load_approved_titles.ipynb <==
# Load Approved Titles

import os
from glob import glob
import pandas as pd

import wmfdata as wmf
from wmfdata import charting, mariadb, hive
from wmfdata.utils import pct_str, pd_display_all
from urllib.parse import unquote

### Read Indonesia data

import glob
import pandas as pd

files = glob.glob('../../data/raw/articles/2019/Indonesia/*.csv')
dfs = []
for file in files:
    df = pd.read_csv(file)
    df['filename'] = file
    dfs.append(df)
articles_raw = pd.concat(dfs, ignore_index=True)

articles = articles_raw.copy(deep=True)

articles = articles.rename(columns={'Wiki':'language_name',
                                    'Type of article': 'contest_article_type',
                                    'Page name'  : 'url'  
                       })

#replace strings to comply with db table values
articles = articles.replace({'language_name' : { 'Bahasa Indonesia' : 'Indonesian',
                                                 'Sunda'            : 'Sundanese',
                                                 'Jawa'             : 'Javanese'
                                               }
                            })

#add page title column
articles['page_title'] =  articles['url'].str.split('.wikipedia.org/wiki/').str[1]

#remove typos - trailing comma at end of page_titles
remove_char_list = [',', '__', '_']
articles['page_title'] = articles['page_title'].str.rstrip(',')
articles['page_title'] = articles['page_title'].str.rstrip('_')
articles['page_title'] = articles['page_title'].str.rstrip('__')




#ensure article links and article info link
articles['filename'] =  articles['filename'].str.split('../../data/raw/articles/2019/Indonesia/').str[1]
articles['url_base'] =  articles['url'].str.split('https://').str[1]
articles['url_base'] = articles['url_base'].str.split('/').str[0]
articles['url_article_info'] = 'https://xtools.wmflabs.org/articleinfo/'+articles['url_base']+'/'+articles['page_title']

df = articles[['page_title', 'language_name', 'contest_article_type', 'filename', 'url', 'url_article_info']]

len(df[df['page_title'].str.contains('%',na=False)])

decoded_df = df[df['page_title'].str.contains('%',na=False)]
df = df[~df['page_title'].str.contains('%',na=False)]

### Read India data

india_filename = '../../data/raw/articles/2019/Tiger 2.0 - article page stats.xlsx'


cols2skip = [0,2,3]
cols = [i for i in range(10) if i not in cols2skip]
dfs = pd.read_excel(indonesia_filename, sheet_name=None,  header=None, usecols=cols) #skiprows=1

df = pd.concat([df.assign(name=n) for n,df in dfs.items()], sort=True)
#df.columns = ['page_title', 'language_name']

#check the langauage names (aka sheets by wiki) and make sure we can identify each language
#pnbwiki	Western Punjabi	https://pnb.wikipedia.org/
#pawiki	Punjabi	Punjabi(Gurumukhi)	https://pa.wikipedia.org/

df['language_name'] = df['language_name'].replace('Bengali', 'Bangla')
df['language_name'] = df['language_name'].replace('Punjabi(Gurumukhi)', 'Punjabi')

### Pull data

#get database_code and language_code, confirm language_code (if needed)
lang_names =tuple(df['language_name'].unique())

ci = wmf.hive.run("""
SELECT  language_code, database_code, language_name
FROM canonical_data.wikis
WHERE language_name IN {lang_names} AND database_group = 'wikipedia'
""".format(lang_names=lang_names))

#merge
df_ci = df.merge(ci, how="left", on=['language_name'])

df_ci_articles = df_ci.loc[df_ci['page_title'].notnull(), ['page_title', 
                                                           'language_name', 
                                                           'database_code', 
                                                           'language_code',
                                                           'filename',
                                                           'contest_article_type', 
                                                           'url',
                                                           'url_article_info'
                                                          ]]

#normalize the titles
df_ci_articles['page_title'] = df_ci_articles['page_title'].str.replace(' ', '_').copy(deep=False)

### Save df

df_ci_articles.to_csv("../../data/raw/articles/2019/clean/articles.csv", sep=',', encoding = 'utf-8', index=False)

df_ci_articles.to_csv("../../data/raw/articles/2019/Indonesia/compiled/articles.csv", sep=',', encoding = 'utf-8', index=False)
