==> scripts/data_wrangling/3d_load_titles_hive.ipynb <==
import pandas as pd
from wmfdata import hive, charting
from wmfdata.charting import comma_fmt, pct_fmt
from wmfdata.utils import pct_str

## Read and clean data

#final interim frame (with new and stub articles)
interim_frame_updated = pd.read_csv("../../data/processed/query_results/content_quality/interim_frame_updated.csv", sep=',', encoding = 'utf-8', parse_dates=True) 

#rename column
interim_frame_updated.rename(columns={'wikicode':'database_code'}, inplace = True)

interim_frame_updated_to_hive = interim_frame_updated[['database_code','article_type','page_id', 'page_title']]

#count num of non zero values
interim_frame_updated_to_hive.astype(bool).sum(axis=0)

#id zeros
interim_frame_updated_to_hive.loc[interim_frame_updated_to_hive['page_id']==0]

#count nans
interim_frame_updated_to_hive.isna().sum()

#remove null page_id rows
interim_frame_updated_to_hive = interim_frame_updated_to_hive.dropna(axis=0, subset=['page_id'])

interim_frame_updated_to_hive.isnull().sum().sum()

## Preparing to create table and load into hive

interim_frame_updated_to_hive.to_csv("../../data/processed/query_results/content_quality/2_interim_frame_updated_to_hive.csv", sep=',', encoding = 'utf-8', index=False)

#truncate #!hive -e "TRUNCATE TABLE IF EXISTS florez.glow_tiger_articles"
!hive -e "Drop TABLE IF EXISTS florez.glow_tiger_articles"

articles = "../../data/processed/query_results/content_quality/2_interim_frame_updated_to_hive.csv"
hive.load_csv(
    articles,
    field_spec="database_code string, article_type string, page_id int, page_title string",
    db_name="florez",
    table_name="glow_tiger_articles",
    sep=','
)
