==> scripts/data_wrangling/3f_topic_handling.ipynb <==
import pandas as pd
import numpy as np
import json

## Final Frame articles

#read articles
#articles = pd.read_csv("../../data/processed/query_results/content_quality/b3_final_frame_updated_pv.csv", sep=',', encoding = 'utf-8')

articles = pd.read_csv("../../data/processed/query_results/content_quality/indonesia/CQ_all_articles.csv")

#create new df from articles which has only pagetitle (underscores not spaces) and QID for each GLOW article
df = articles.filter(['QID', 'page_title'], axis=1)
#df['QID'] = df['QID'].astype(int)
#df['QID'] = 'Q'+ df['QID'].astype(str)

len(df.loc[df['QID']==0])

#remove Q0 ids which will block the model from working
#Qzero = df.loc[df['QID']=='Q0']
df = df[df.QID !=0]

#df['QID'] = df['QID'].apply(str)
df['QID'] = 'Q' + df['QID'].astype(str)
df.rename(columns={'page_title':'title'}, inplace=True)

### Topics

#### data to json .txt file

#QID string item, article item
f = open("../../data/processed/query_results/topic_lists/topic_model/inputs_GLOW_Saraswati1.txt", "w")

for row in df.iterrows():
    row[1].to_json(f)
    f.write("\n")

#### >> get topic data, using model & terminal on local 

#https://github.com/geohci/wikidata-topic-model
#https://meta.wikimedia.org/wiki/Research_talk:Characterizing_Wikipedia_Reader_Behaviour/Demographics_and_Wikipedia_use_cases/Work_log/2019-09-11
#start terminal
>> $ source activate fasttext
>> $ cd <appropriate folder path>
>> $ python3 wikidata_ids_to_topics_api.py


##### read results

# import the json objects from the .txt file into a list of dictionaries
#with open('../../data/processed/query_results/topic_lists/topic_model/output_GLOW_ProjectTiger2topic_predictions.txt') as file:
with open('../../data/processed/query_results/topic_lists/topic_model/Indonesia_outputs/output_GLOW_Saraswati1topic_predictions.txt') as file:

    Q_articles = []
    for line in file:
        Q_articles.append(json.loads(line))
        
Q_articles_df = pd.DataFrame(Q_articles)

Q_articles_df.info()

# https://stackoverflow.com/questions/18665284/how-do-i-access-embedded-json-objects-in-a-pandas-dataframe
##seperate out the labels column to take a closer look
labels_df = Q_articles_df['labels'].apply(pd.Series)

#pull out the label % confidence number from label 1 
label1_confnum = labels_df[0].apply(pd.Series)
#pull out the label % confidence number from label2
label2_confnum = labels_df[1].apply(pd.Series)

#rename column names
label1_confnum.rename(columns={0:'predicted_label_1', 1:'pl1_score'},inplace=True)
label2_confnum.rename(columns={0:'predicted_label_2', 1:'pl2_score'},inplace=True)

#concat back into the df
labels_confnum_combo = pd.concat([label1_confnum, label2_confnum.apply(pd.Series)], axis=1)
articles_labels_df = pd.concat([Q_articles_df, labels_confnum_combo.apply(pd.Series)], axis=1)

#chage QID to q_id
#articles_labels_df.rename(columns={'QID':'q_id'},inplace=True)
articles_labels_df.rename(columns={'title':'page_title', 'labels': 'predicted_labels'},inplace=True)

articles_labels_df.info()

#save to df
articles_labels_df.to_csv("../../data/processed/query_results/topic_lists/GLOW_Saraswati1_topics_df.csv", sep=',', encoding = 'utf-8', index=False)

#prep the merge, if necessary
#interim_articles['QID'] = interim_articles['QID'].astype(int)
#interim_articles['QID'] = 'Q'+ interim_articles['QID'].astype(str)
#merge
interim_frame_updated_pv_topics = pd.merge(interim_articles, articles_int_labels_df, on=['page_title', 'QID'], how="left").fillna(0);
#save to csv
interim_frame_updated_pv_topics.to_csv("../../data/processed/query_results/content_quality/4_interim_frame_updated_pv_topcs.csv", sep=',', encoding = 'utf-8', index=False)


#### wikidata P31 value -- at present requires additional optimization to run and provides few results given our inputs

preptitles = ', '.join('"{0}"@pa'.format(w) for w in clean_new_pagetitles_list) 

query = """
SELECT DISTINCT ?item ?lemma ?instance_of 
WHERE {
  ?item wdt:P31 ?instance_of.    
  ?sitelink schema:about ?item;
   schema:isPartOf <https://pa.wikipedia.org/>;
   schema:name ?lemma.
   FILTER (?lemma IN (  %s )).
   SERVICE wikibase:label { bd:serviceParam wikibase:language "en". }
}   
"""%preptitles

print(query)

# read manual P31 pull data
PA_P31 = pd.read_csv("../GlowPrivate/data/raw/P31/PA_P31data.csv", sep=',', encoding = 'utf-8')

# https://stackoverflow.com/questions/13650840/use-another-sparql-inside-sparql-in-clause
# https://github.com/santhoshtr/wikivue/blob/master/src/wiki/api/wikidata.js
# P31 matched to top topics QUERY - 

SELECT DISTINCT ?lemma ?categoryLabel ?item ?itemLabel ?itemDescription 
    WHERE{
      VALUES ?category {
        wd:Q735
        wd:Q2095
        wd:Q309
        wd:Q178561
        wd:Q31629
        wd:Q1071
        wd:Q192864
        wd:Q395
        wd:Q11016
        wd:Q7189713
        wd:Q1322005
        wd:Q315
      } #
      ?item p:P31/ps:P31/wdt:P279* ?category. # Item is of type or subtype an item in the above list of categories
      ?article schema:about ?item ; 
               schema:isPartOf <https://{%s}.wikipedia.org/>. # Exist in ${language}
      FILTER (?lemma IN ({%s})).
      SERVICE wikibase:label { bd:serviceParam wikibase:language "en". }
    }

# https://stackoverflow.com/questions/13650840/use-another-sparql-inside-sparql-in-clause
# https://github.com/santhoshtr/wikivue/blob/master/src/wiki/api/wikidata.js

import requests
import pandas as pd
from collections import OrderedDict
from ratelimit import limits, sleep_and_retry
        

ONE_MINUTE = 60
@sleep_and_retry
@limits(calls=15, period=ONE_MINUTE)
def call_api():
    
    url = 'https://query.wikidata.org/sparql'
    headers = {'Content-Type': 'application/sparql-query', 
               'User-Agent': 'Partnerships GLOW project Data Analyst'}
    query = """
            SELECT DISTINCT ?item ?lemma ?instance_of 
            WHERE {
              ?item wdt:P31 ?instance_of.    
              ?sitelink schema:about ?item;
               schema:isPartOf <https://en.wikipedia.org/>;
               schema:name ?lemma.
               FILTER (?lemma IN (  "Sam Walton"@en, "Nutcracker"@en ) ).
               SERVICE wikibase:label { bd:serviceParam wikibase:language "en". }
            }   
            """
    r = requests.post(url, headers=headers, data=query)
    if r.status_code != 200:
        raise Exception('API response: {}'.format(r.status_code))
    return r

call_api()

data = r.json()

gwd = []
for item in data['results']['bindings']:
    gwd.append(OrderedDict({
        'page_title': item['lemma']['value'],
        'q_id': item['item']['value'],
        'instance_of': item['instance_of']['value'],
        }))

q_df = pd.DataFrame(gwd)
q_df.set_index('page_title', inplace=True)
q_df = q_df.rename(columns={'q_id':'q_url', 'instance_of':'P31_url'})
#q_df['qid'] = q_df['q_id_url'].str.extract('([Q]\d+|1)', expand=True)# includes the Q at the start
q_df['qid'] = q_df['q_url'].str.extract('([0-9]+(?=[^Q]*$))', expand=True) # excludes the Q at the start
q_df['P31'] = q_df['P31_url'].str.extract('([Q]\d+|1)', expand=True)

#P31 processing
# see: https://github.com/santhoshtr/wikivue/blob/master/src/wiki/topics.json

lang_top_topics = [
    {
        "wd": "Q735",
        "name": "Arts",
        "image": "https://commons.wikimedia.org/wiki/Special:FilePath/Sebastiano_Ricci_002.jpg"
    },
    {
        "wd": "Q2095",
        "name": "Food",
        "image": "https://commons.wikimedia.org/wiki/Special:FilePath/Good_Food_Display_-_NCI_Visuals_Online.jpg"
    },
    {
        "wd": "Q309",
        "name": "History",
        "image": "https://commons.wikimedia.org/wiki/Special:FilePath/Ancientlibraryalex.jpg"
    },
    {
        "wd": "Q178561",
        "name": "Battles",
        "image": "https://commons.wikimedia.org/wiki/Special:FilePath/Battle_of_Gibraltar_1607.jpg"
    },
    {
        "wd": "Q31629",
        "name": "Sports",
        "image":"https://commons.wikimedia.org/wiki/Special:FilePath/Youth-soccer-indiana.jpg"
    },
    {
        "wd": "Q1071",
        "name": "Geography",
        "image": "https://commons.wikimedia.org/wiki/Special:FilePath/OrteliusWorldMap.jpeg"
    },
    {
        "wd": "Q192864",
        "name": "Science",
        "image": "https://commons.wikimedia.org/wiki/Special:FilePath/CMS_Higgs-event.jpg"
    },
    {
        "wd": "Q395",
        "name": "Maths",
        "image": "https://commons.wikimedia.org/wiki/Special:FilePath/CMAP_-_Centre_de_Math%C3%A9matiques_Appliqu%C3%A9es_de_l%27Ecole_polytechnique_%2825911635211%29.jpg"
    },
    {
        "wd": "Q11016",
        "name": "Tech",
        "image": "https://commons.wikimedia.org/wiki/Special:FilePath/Technologie.jpg"
    },
    {
        "wd":"Q7189713",
        "name": "Health",
        "image": "http://commons.wikimedia.org/wiki/Special:FilePath/Heart_anterior_exterior_view.jpg"
    },
    {
        "wd":"Q1322005",
        "name": "Nature",
        "image": "http://commons.wikimedia.org/wiki/Special:FilePath/Desert_Electric.jpg"
    },
    {
        "wd":"Q315",
        "name": "Language",
        "image": "http://commons.wikimedia.org/wiki/Special:FilePath/File:Pieter_Bruegel_the_Elder_-_The_Tower_of_Babel_(Vienna)_-_Google_Art_Project_-_edited.jpg"
    }
]

for d in lang_top_topics:
    print(d['wd'])

#q_df['qid'] = q_df['q_id_url'].str.extract('([Q]\d+|1)', expand=True)# includes the Q at the start
q_df['qid'] = q_df['q_url'].str.extract('([0-9]+(?=[^Q]*$))', expand=True) # excludes the Q at the start
q_df['P31'] = q_df['P31_url'].str.extract('([Q]\d+|1)', expand=True)
