==> scripts/data_wrangling/6c_collect_baseline_contest_region.ipynb <==
# GLOW - Setup: Collect Baselines - contest region

## Table of Contents  <a class="anchor" id="toc"></a>

* [GLOW Wiki Baselines](#top)

    1. [Devices](#devices)
    2. [Editors](#editors)
    3. [Articles](#articles)


## Devices <a class="anchor" id="article_detail"></a>
[Back to Table of Contents](#toc)

#### Monthly Unique Devices <a class="anchor" id="devices"></a>
[Back to Table of Contents](#toc)

mca_uds_contest_r = '''
SELECT
  SUM(uniques_estimate) / 5 AS monthly_unique_devices
FROM wmf.unique_devices_per_domain_monthly
WHERE 
    CONCAT(year,LPAD(month,2,'0')) >= ({contest_start_dt_pv})
    AND CONCAT(year,LPAD(month,2,'0')) < ({contest_end_dt_next_m_pv})
    AND country_code IN ('IN')
    AND domain IN ({india_domains})
''' 

## Editors<a class="anchor" id="editors"></a>
[Back to Table of Contents](#toc)

#https://hue.wikimedia.org/hue/metastore/table/cchen/editor_month

#### Monthly editors & monthly new <a class="anchor" id="editors_monthly"></a>
[Back to Table of Contents](#toc)

#adapted from:
#https://github.com/wikimedia-research/wiki-segmentation
#https://github.com/wikimedia-research/Editing-movement-metrics

mce_contest_r = '''
SELECT
    COUNT(*) / 5 AS monthly_editors,
    sum(CAST(TRUNC(em.user_registration, 'MM') = TRUNC(month, 'MM') AS INT))/ 5 AS monthly_new_editors
FROM cchen.editor_month em
WHERE
    em.month >= "{contest_start_dt}"
    AND em.month < "{contest_end_dt_next_m}"
    AND em.wiki IN ({india_wiki_dbs}) 
    AND em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot"))
'''

#### Monthly New Active Editors & monthly active editors <a class="anchor" id="editors_active"></a>
[Back to Table of Contents](#toc)

#monthly active editors
#adapted from:
#https://github.com/wikimedia-research/wiki-segmentation
#https://github.com/wikimedia-research/Editing-movement-metrics

mnae_contest_r = '''
SELECT
    COUNT(*) / 5 AS monthly_active_editors,
    SUM(
        CAST(TRUNC(em.user_registration, 'MM') = TRUNC(month, 'MM') AS INT)
        )/ 5 AS monthly_new_active_editors
FROM cchen.editor_month em
WHERE
    content_edits >= 5 
    AND em.month >= "{contest_start_dt}"
    AND em.month < "{contest_end_dt_next_m}"
    AND em.wiki IN ({india_wiki_dbs}) 
    AND em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot"))  
'''

#### Monthly Editors - including group of big wikis <a class="anchor" id="editors_active"></a>
[Back to Table of Contents](#toc)

mae_contest_r = '''
SELECT
    COUNT(*) / 5 AS indic_editors_on_big_wikis_m
FROM cchen.editor_month em
WHERE
    em.month >= "{contest_start_dt}"
    AND em.month < "{contest_end_dt_next_m}"
    AND em.wiki IN {wikis_big} 
    AND em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot"))  
'''

#### New editor retention <a class="anchor" id="new_editor_retention"></a>
[Back to Table of Contents](#toc)

#adapted from:
#https://github.com/wikimedia-research/wiki-segmentation
#https://github.com/wikimedia-research/Editing-movement-metrics

ner_contest_r = '''
select 
    sum(cast(2nd_month_edits >= 1 as int)) / sum(cast(1st_month_edits >= 1 as int)) AS new_editor_retention
from neilpquinn.new_editors
where 
    cohort >= "{contest_start_dt_12m_pre}"
    AND cohort < "{TODAY_DATE}"
    AND wiki IN ({india_wiki_dbs})
'''

with open("queries/new_editor_retention.hql") as f:
    q = f.read()

ner = wmf.hive.run(
    q.format(start = "{contest_start_dt_12m_pre}", end = "{TODAY_DATE}").format(**query_vars))

##ner = x where x in IN ({india_wiki_dbs})

## Articles<a class="anchor" id="editors"></a>
[Back to Table of Contents](#toc)

#### New articles <a class="anchor" id="new_articles"></a>
[Back to Table of Contents](#toc)

#adapted from https://github.com/wikimedia-research/2018-19-Language-annual-plan-metrics/blob/master/Language-metrics.ipynb
#https://wikitech.wikimedia.org/wiki/Analytics/Data_Lake/Edits/Mediawiki_history
#revision	create = Making an edit
# 6 months ≈ 26 weeks = 252 days
# period below starts 2019/07 

# Making the first edit to a page

m_new_article_counts_contest_r = ''' 
select
    count(*)/5 AS mon_new_articles
from wmf.mediawiki_history mh
left join event_sanitized.serversideaccountcreation ssac
on
    ssac.event.username = event_user_text and
    ssac.year >= 0
where
    mh.snapshot = "{MWH_SNAPSHOT}"
    AND mh.event_timestamp >= "{contest_start_dt_FULL}"
    AND mh.event_timestamp < "{contest_end_dt_FULL}" 
    AND event_entity = "revision"
    AND event_type = "create"
    AND mh.wiki_db in ({india_wiki_dbs})
''' 

#### avg_num_new_articles_edited <a class="anchor" id="new_articles"></a>
[Back to Table of Contents](#toc)

#adapted from https://github.com/wikimedia-research/2018-19-Language-annual-plan-metrics/blob/master/Language-metrics.ipynb
#https://wikitech.wikimedia.org/wiki/Analytics/Data_Lake/Edits/Mediawiki_history
#page	create = Making the first edit to a page

#6 months ≈ 26 weeks = 252 days

#Making the first edit to a page
m_new_articles_edited_contest_r = ''' 
SELECT
    count(*)/5 AS mon_new_articles_edited
FROM wmf.mediawiki_history mh
WHERE
    mh.snapshot = "{MWH_SNAPSHOT}" 
    AND mh.event_timestamp >= "{contest_start_dt_FULL}"
    AND mh.event_timestamp < "{contest_end_dt_FULL}" 
    AND event_entity = "page"
    AND event_type = "create"
    AND wiki_db in ({india_wiki_dbs})
''' 

#AND ssac.webhost LIKE '%wikipedia.org'

#### Existing articles, recently edited <a class="anchor" id="article_edits"></a>
[Back to Table of Contents](#toc)

#adapted from https://github.com/wikimedia-research/2018-19-Language-annual-plan-metrics/blob/master/Language-metrics.ipynb
#https://wikitech.wikimedia.org/wiki/Analytics/Data_Lake/Edits/Mediawiki_history
#6 months ≈ 26 weeks = 252 days
#period below starts 2019/07 

eae_contest_r = ''' 
select
    count(*)/5 as avg_n_existing_articles_edited
from wmf.mediawiki_history mh
where
    mh.snapshot = "{MWH_SNAPSHOT}"  
    AND mh.event_timestamp >= "{contest_start_dt_FULL}"
    AND mh.event_timestamp < "{contest_end_dt_FULL}" 
    AND event_entity = "revision"
    AND event_type = "create" 
    AND wiki_db in ({india_wiki_dbs})
''' 

#### Daily revisions by wiki <a class="anchor" id="daily_wiki_revisions"></a>
[Back to Table of Contents](#toc)

##Daily revisions by wiki

#`dr` stands for "daily revisions"
dra_contest_r = ''' 
    select
        sum(if(metric = "daily_edits", value, 0)) - sum(if(metric = "daily_edits_by_bot_users", value, 0))/12 as nonbot_revs
    from wmf.mediawiki_metrics
    where
        snapshot = "{MWH_SNAPSHOT}" 
        AND dt >="{contest_start_dt_FULL}"
        AND dt <"{contest_end_dt_FULL}"
        AND metric in ("daily_edits", "daily_edits_by_bot_users")
        AND wiki_db IN ({india_wiki_dbs})
'''
