==> scripts/data_wrangling/7a_process_baseline_monthly_wikis.ipynb <==
# process/get baselines - monthly wikis

from pyspark.sql.types import ArrayType, StringType

from pyspark.context import SparkContext
from pyspark.sql.session import SparkSession

import datetime as dt

import pandas as pd
import numpy as np

%run 2b_data_handling.ipynb
%run 6a_collect_baseline_monthly_wikis.ipynb

countries = spark.sql(countries_r.format(**query_vars)).toPandas();

wikis = spark.sql(wikis_r.format(**query_vars)).toPandas()

wikis.to_csv("../../data/processed/query_results/wikis.csv", sep=',', encoding = 'utf-8', index=False)

wikis_c = wikis.copy()
wikis_c = wikis_c.reset_index()
india_wikis = wikis_c.loc[wikis_a['database_code'].isin(india_glow_wiki_dbs_mariadb)]

wikis = pd.read_csv("../../data/processed/query_results/wikis.csv", sep=',', encoding = 'utf-8')

#### Devices <a class="anchor" id="article_detail"></a>
[Back to Table of Contents](#toc)

mca_uds = spark.sql(mca_uds_r.format(**query_vars)).toPandas()
mca_uds["domain_name"] = "https://" + mca_uds["domain_name"]

wikis = wikis.merge(mca_uds, how='left', on='domain_name')

#### Editors<a class="anchor" id="editors"></a>
[Back to Table of Contents](#toc)

mce = spark.sql(mce_r.format(**query_vars)).toPandas()
merge_in(mce)

mnae = spark.sql(mnae_r.format(**query_vars)).toPandas()
merge_in(mnae)

mae = spark.sql(mae_r.format(**query_vars)).toPandas()
merge_in(mnae)

ner = spark.sql(ner_r.format(**query_vars)).toPandas()
merge_in(ner)

#### Articles<a class="anchor" id="editors"></a>
[Back to Table of Contents](#toc)

m_new_article_counts = spark.sql(m_new_article_counts_r.format(**query_vars)).toPandas()
merge_in(m_new_article_counts)

m_new_articles_edited = spark.sql(m_new_articles_edited_r.format(**query_vars)).toPandas()
merge_in(m_new_articles_edited)

eae = spark.sql(eae_r.format(**query_vars)).toPandas()
merge_in(eae)

dra = spark.sql(dra_r.format(**query_vars)).toPandas()
merge_in(dra)

wikis

## save to csv

wikis.to_csv("../../data/processed/query_results/regional_counts/baseline/baseline_monthly_wikis.csv", sep=',', encoding = 'utf-8', index=False)
