==> scripts/data_wrangling/7c_process_baseline_contest_region.ipynb <==
# process/get contest period - region

from pyspark.sql.types import ArrayType, StringType

from pyspark.context import SparkContext
from pyspark.sql.session import SparkSession

import datetime as dt

import pandas as pd
import numpy as np

%run 2b_data_handling.ipynb
%run 6c_collect_baseline_contest_region.ipynb

#### Devices <a class="anchor" id="article_detail"></a>
[Back to Table of Contents](#toc)

mca_uds = spark.sql(mca_uds_contest_r.format(**query_vars)).toPandas()

#### Editors<a class="anchor" id="editors"></a>
[Back to Table of Contents](#toc)

mce = spark.sql(mce_contest_r.format(**query_vars)).toPandas()

mnae = spark.sql(mnae_contest_r.format(**query_vars)).toPandas()

mae = spark.sql(mae_contest_r.format(**query_vars)).toPandas()

ner = spark.sql(ner_contest_r.format(**query_vars)).toPandas()

#### Articles<a class="anchor" id="editors"></a>
[Back to Table of Contents](#toc)

m_new_article_counts = spark.sql(m_new_article_counts_contest_r.format(**query_vars)).toPandas()

m_new_articles_edited = spark.sql(m_new_articles_edited_contest_r.format(**query_vars)).toPandas()

eae = spark.sql(eae_contest_r.format(**query_vars)).toPandas()

dra = spark.sql(dra_contest_r.format(**query_vars)).toPandas()

## unified df

contest_dfs = [mca_uds, ner, mce, mnae, mae, 
                                  m_new_article_counts, m_new_articles_edited,
                                  eae, dra]
contest_period_region = pd.concat(contest_dfs, axis=1)

contest_period_region

## save to csv

contest_period_region.to_csv("../../data/processed/query_results/regional_counts/baseline/contest_period_region.csv", sep=',', encoding = 'utf-8', index=False)
