==> scripts/data_wrangling/8a_collect_contest_editors_region.ipynb <==
# Collect contest editors - region

## Table of Contents  <a class="anchor" id="toc"></a>

* [GLOW Wiki Baselines](#top)

    1. [Editors](#editors)


## Editors<a class="anchor" id="editors"></a>
[Back to Table of Contents](#toc)

#https://hue.wikimedia.org/hue/metastore/table/cchen/editor_month

#### Monthly editors & monthly new <a class="anchor" id="editors_monthly"></a>
[Back to Table of Contents](#toc)

#adapted from:
#https://github.com/wikimedia-research/wiki-segmentation
#https://github.com/wikimedia-research/Editing-movement-metrics

mce_contest_r = '''
SELECT
    COUNT(*) / 5 AS monthly_editors,
    sum(CAST(TRUNC(em.user_registration, 'MM') = TRUNC(month, 'MM') AS INT))/ 5 AS monthly_new_editors
FROM florez.glow_tiger_participants gtp  
LEFT JOIN cchen.editor_month em
    ON gtp.username = em.user_name 
WHERE
    em.month >= "{contest_start_dt}"
    AND em.month < "{contest_end_dt_next_m}"
    AND em.wiki IN ({india_wiki_dbs}) 
    AND em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot"))
'''

#adapted from:
#https://github.com/wikimedia-research/wiki-segmentation
#https://github.com/wikimedia-research/Editing-movement-metrics

mce_pre_r = '''
SELECT
    COUNT(*) / 12 AS monthly_editors,
    sum(CAST(TRUNC(em.user_registration, 'MM') = TRUNC(month, 'MM') AS INT))/ 12 AS monthly_new_editors
FROM florez.glow_tiger_participants gtp  
LEFT JOIN cchen.editor_month em
    ON gtp.username = em.user_name 
WHERE
    em.month >= "{contest_start_dt_12m_pre}"
    AND em.month < "{contest_start_dt}"
    AND em.wiki IN ({india_wiki_dbs}) 
    AND em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot"))
'''

#### Monthly active editors <a class="anchor" id="editors_active"></a>
[Back to Table of Contents](#toc)

#monthly active editors
#adapted from:
#https://github.com/wikimedia-research/wiki-segmentation
#https://github.com/wikimedia-research/Editing-movement-metrics

mnae_contest_r = '''
SELECT
    COUNT(*) / 5 AS monthly_active_editors,
    SUM(
        CAST(TRUNC(em.user_registration, 'MM') = TRUNC(month, 'MM') AS INT)
        )/ 5 AS monthly_new_active_editors
FROM florez.glow_tiger_participants gtp  
LEFT JOIN cchen.editor_month em
    ON gtp.username = em.user_name 
WHERE
    content_edits >= 5 
    AND em.month >= "{contest_start_dt}"
    AND em.month < "{contest_end_dt_next_m}"
    AND em.wiki IN ({india_wiki_dbs}) 
    AND em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot"))  
'''

mnae_pre_r = '''
SELECT
    COUNT(*) / 12 AS monthly_active_editors,
    SUM(
        CAST(TRUNC(em.user_registration, 'MM') = TRUNC(month, 'MM') AS INT)
        )/ 12 AS monthly_new_active_editors
FROM florez.glow_tiger_participants gtp  
LEFT JOIN cchen.editor_month em
    ON gtp.username = em.user_name 
WHERE
    content_edits >= 5 
    AND em.month >= "{contest_start_dt_12m_pre}"
    AND em.month < "{contest_start_dt}"
    AND em.wiki IN ({india_wiki_dbs}) 
    AND em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot"))  
'''

#### Monthly Editors - including group of big wikis <a class="anchor" id="editors_active"></a>
[Back to Table of Contents](#toc)

mae_contest_r = '''
SELECT
    COUNT(*) / 5 AS indic_editors_on_big_wikis_m
FROM florez.glow_tiger_participants gtp  
LEFT JOIN cchen.editor_month em
    ON gtp.username = em.user_name 
WHERE
    em.month >= "{contest_start_dt}"
    AND em.month < "{contest_end_dt_next_m}"
    AND em.wiki IN {wikis_big} 
    AND em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot"))  
'''

mae_pre_r = '''
SELECT
    COUNT(*) / 12 AS indic_editors_on_big_wikis_m
FROM florez.glow_tiger_participants gtp  
LEFT JOIN cchen.editor_month em
    ON gtp.username = em.user_name 
WHERE
    em.month >= "{contest_start_dt_12m_pre}"
    AND em.month < "{contest_start_dt}"
    AND em.wiki IN {wikis_big} 
    AND em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot"))  
'''

#adapted from:
#https://github.com/wikimedia-research/wiki-segmentation
#https://github.com/wikimedia-research/Editing-movement-metrics

ner_contest_r = '''
select 
    sum(cast(ne.2nd_month_edits >= 1 as int)) / sum(cast(ne.1st_month_edits >= 1 as int)) AS new_editor_retention
FROM florez.glow_tiger_participants gtp  
LEFT JOIN neilpquinn.new_editors ne
    ON gtp.username = ne.user_name
where 
    ne.cohort >= "{contest_start_dt}"
    AND ne.cohort < "{contest_end_dt_next_m}"
    AND ne.wiki IN ({india_wiki_dbs})
'''

ner_pre_r = '''
select 
    sum(cast(ne.2nd_month_edits >= 1 as int)) / sum(cast(ne.1st_month_edits >= 1 as int)) AS new_editor_retention
FROM florez.glow_tiger_participants gtp  
LEFT JOIN neilpquinn.new_editors ne
    ON gtp.username = ne.user_name
where 
    ne.cohort >= "{contest_start_dt_12m_pre}"
    AND ne.cohort < "{contest_start_dt}"
    AND ne.wiki IN ({india_wiki_dbs})
'''

### Editor's registration dates, to identify new editors

#event_user_creation_timestamp	string	Creation timestamp of the user that caused the event (from logging table)
#https://wikitech.wikimedia.org/wiki/Analytics/Data_Lake/Edits/MediaWiki_history#Schema

editor_reg_contest_2019_r = '''
SELECT
    gtp.username,
    em.wiki,
    TRUNC(em.user_registration, 'MM') AS reg_date
FROM florez.glow_tiger_participants gtp  
LEFT JOIN cchen.editor_month em
    ON gtp.username = em.user_name 
WHERE 
    em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot")) 
'''

editor_reg_contest_2018_r = '''
SELECT
    ptp.username,
    em.wiki,
    TRUNC(em.user_registration, 'MM') AS reg_date
FROM florez.glow_tiger_pilot_participants ptp  
LEFT JOIN cchen.editor_month em
    ON ptp.username = em.user_name 
WHERE 
    em.user_id != 0 
    AND em.bot_by_group = FALSE
    AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot")) 
'''

### Edits

GLOW_editor_edits_r = '''
SELECT mh.event_user_text, 
       mh.event_timestamp
FROM florez.glow_tiger_participants gtp
JOIN wmf.mediawiki_history mh
    on gtp.username = mh.event_user_text
WHERE (mh.snapshot='2020-03') 
    AND mh.event_timestamp >= '2018-01-01' 
    AND event_entity = "revision"
    AND event_type = "create"
    --order by event_user_revision_count desc
'''

weekly_edits_query_ptp_grantees_r='''
SELECT mh.event_user_text, 
       WEEKOFYEAR(mh.event_timestamp) AS week_n,
       COUNT(mh.revision_id) AS edits
FROM florez.glow_tiger_pilot_participants ptp
JOIN wmf.mediawiki_history mh
    on ptp.username = mh.event_user_text
WHERE (mh.snapshot='2020-03') 
    AND mh.event_timestamp >= '2018-01-01' 
    AND mh.event_timestamp < '2019-01-01'
    AND event_entity = "revision"
    AND event_type = "create"
    AND ptp.incentive_recipient = True
    AND ptp.incentive_type = 'laptop'
    AND mh.revision_minor_edit = false
GROUP BY mh.event_user_text, weekofyear(mh.event_timestamp)
'''

edits_query_ptp_grantees_r='''
SELECT mh.event_user_text, 
       mh.event_timestamp AS edit_date,
       COUNT(mh.revision_id) AS edits
FROM florez.glow_tiger_pilot_participants ptp
JOIN wmf.mediawiki_history mh
    on ptp.username = mh.event_user_text
WHERE (mh.snapshot='2020-03') 
    AND mh.event_timestamp >= '2017-01-01' 
    AND mh.event_timestamp < '2020-04-01'
    AND event_entity = "revision"
    AND event_type = "create"
    AND ptp.incentive_recipient = True
    AND ptp.incentive_type = 'laptop'
    AND mh.revision_minor_edit = false
GROUP BY mh.event_user_text, mh.event_timestamp
'''

weekly_edits_query='''
SELECT gtp.username,
    WEEKOFYEAR(event_timestamp) AS week_n, 
    COUNT(revision_id) AS edits
FROM florez.glow_tiger_participants gtp
JOIN wmf.mediawiki_history mh
    on gtp.username = mh.event_user_text
WHERE
    event_entity = 'revision' 
    AND event_type = 'create' 
    AND DATE(event_timestamp) >= '2018-01-01'  
    AND snapshot = '2020-04'
    AND incentive_recipient = True
    AND incentive_type = 'laptop'
GROUP BY weekofyear(event_timestamp) 
'''

weekly_edits_query_ptp='''
SELECT gtp.username,
    WEEKOFYEAR(mh.event_timestamp) AS week_n, 
    COUNT(mh.revision_id) AS edits
FROM florez.glow_tiger_pilot_participants gtp
JOIN wmf.mediawiki_history mh
    on gtp.username = mh.event_user_text
WHERE (mh.snapshot='2020-04')
    AND mh.event_entity = 'revision' 
    AND mh.event_type = 'create' 
    AND mh.event_timestamp >= '2018-01-01'  
    AND gtp.incentive_recipient = True
    AND gtp.incentive_type = 'laptop'
    AND mh.revision_minor_edit = false
GROUP BY gtp.username, mh.event_timestamp
'''

weekly_edits_query_ptp='''
SELECT gtp.username,
    WEEKOFYEAR(mh.event_timestamp) AS week_n, 
    COUNT(mh.revision_id) AS edits
FROM florez.glow_tiger_pilot_participants gtp
JOIN wmf.mediawiki_history mh
    on gtp.username = mh.event_user_text
WHERE (mh.snapshot='2020-04')
    AND mh.event_entity = 'revision' 
    AND mh.event_type = 'create' 
    AND DATE(event_timestamp) >= '2018-01-01'  
    AND gtp.incentive_recipient = True
    AND gtp.incentive_type = 'laptop'
GROUP BY gtp.username, mh.event_timestamp
'''

import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
import matplotlib.dates as mdates
import seaborn as sns
import pandas as pd
import numpy as np

import wmfdata as wmf
from wmfdata import charting, mariadb, hive, spark
from wmfdata.utils import pct_str, pd_display_all
#%load_ext sql_magic
%matplotlib inline

%run .././data_wrangling/2b_data_handling.ipynb

editors = pd.read_csv("../../data/processed/participants_list_received_clean/indonesia/Saraswati1.csv", sep=',', encoding = 'utf-8')

editor_data = []

def get_editor_data_hive(df):
    

    GLOW_editor_edits_query = '''
    SELECT mh.event_user_text, 
           mh.event_timestamp
    FROM wmf.mediawiki_history mh
    WHERE (mh.snapshot='2020-08') 
        AND mh.event_timestamp >= '2019-05-01' 
        AND event_entity = "revision"
        AND event_type = "create"
        AND wiki_db = '{wiki_db}' 
        AND mh.event_user_text IN {users}
        --order by event_user_revision_count desc
    '''
        
    for wiki in df['database_code'].unique(): 
        grouping = df.loc[df['database_code'] == wiki].groupby('database_code')['user_name'].apply(pd.DataFrame)
        users = tuple(list(grouping[wiki]))
        user = grouping.reset_index(drop=True).iloc[0][0]
        wiki_db = wiki
        editor_results = spark.run(GLOW_editor_edits_query.format(users=users, wiki_db=wiki_db ))
        
        editor_data.append(editor_results)
    return(editor_data)

get_editor_data_hive(editors)
query_results= pd.concat(editor_data)

query_results.info()

query_results['event_timestamp'] = query_results['event_timestamp'].apply(pd.to_datetime)
#query_results['event_timestamp'] =  pd.to_datetime(query_results['event_timestamp'], utc=True)
#query_results['event_timestamp'] = query_results['event_timestamp'].dt.date

id_glow_start = query_vars['contest_start_dt']
id_glow_end = query_vars['contest_end_dt']
id_glow_2019_12M_pre_mask = (query_results['event_timestamp']>= query_vars['contest_start_dt_first_day_12m_pre']) & (query_results['event_timestamp'] < id_glow_start)
id_glow_2019_5M_pre_mask = (query_results['event_timestamp']>= query_vars['contest_start_dt_first_day_5m_pre']) & (query_results['event_timestamp'] < id_glow_start)
id_glow_mask = (query_results['event_timestamp']>= id_glow_start) & (query_results['event_timestamp'] <= id_glow_end) 

id_glow = query_results.loc[id_glow_mask]
id_glow_2019_12M_pre = query_results.loc[id_glow_2019_12M_pre_mask]
id_glow_2019_5M_pre = query_results.loc[id_glow_2019_5M_pre_mask]

id_glow_2019_12M_pre.describe()

id_glow_2019_5M_pre.describe()

id_glow.describe()

df = editors.merge(query_results, how='left', right_on='event_user_text', left_on='user_name')

#df['user_name'].groupby('event_timestamp').agg('count')

df.groupby([df['event_timestamp'].dt.year.rename('year'), df['event_timestamp'].dt.month.rename('month')]).agg({'count'})

df['user_name'].groupby(df.event_timestamp.dt.to_period("M")).agg('count')

df['month_year']= df['event_timestamp'].dt.to_period('M')

counts = df.groupby(['user_name', 'month_year','received_incentive']).size()
#counts.sort_values('month_year', ascending=False)

counts = counts.reset_index().rename(columns={0: 'edit_count'})

count_id_glow_2019_12M_pre_mask = (counts['month_year']>= query_vars['contest_start_dt_first_day_12m_pre']) & (counts['month_year'] < id_glow_start)
count_id_glow_2019_5M_pre_mask = (counts['month_year']>= query_vars['contest_start_dt_first_day_5m_pre']) & (counts['month_year'] < id_glow_start)
count_id_glow_mask = (counts['month_year']>= id_glow_start) & (counts['month_year'] <= id_glow_end) 

count_id_glow = counts.loc[count_id_glow_mask]
count_id_glow_2019_12M_pre = counts.loc[count_id_glow_2019_12M_pre_mask]
count_id_glow_2019_5M_pre = counts.loc[count_id_glow_2019_5M_pre_mask]

count_id_glow_2019_5M_pre.describe()

count_id_glow.describe()

editor_data_signup = []

def get_editor_signup_data_hive(df):
    

    editor_reg_contest_query = '''
    SELECT
        em.user_name,
        em.wiki,
        TRUNC(em.user_registration, 'MM') AS reg_date
    FROM cchen.editor_month em
    WHERE 
        em.user_id != 0
        AND em.user_name IN {users}
        AND em.bot_by_group = FALSE
        AND (em.user_name not regexp "bot\\b" or em.user_name in ("Paucabot", "Niabot", "Marbot")) 
    '''
    
    

        
    for wiki in df['database_code'].unique(): 
        grouping = df.loc[df['database_code'] == wiki].groupby('database_code')['user_name'].apply(pd.DataFrame)
        users = tuple(list(grouping[wiki]))
        user = grouping.reset_index(drop=True).iloc[0][0]
        wiki_db = wiki
        editor_results = spark.run(editor_reg_contest_query.format(users=users, wiki_db=wiki_db ))
        
        editor_data_signup.append(editor_results)
    return(editor_data_signup)

get_editor_signup_data_hive(editors)
query_results_signups = pd.concat(editor_data_signup)

query_results_signups['reg_date'] = query_results_signups['reg_date'].apply(pd.to_datetime)

query_results_signups = query_results_signups.sort_values(by=['user_name', 'reg_date'], ascending=False)

qrs = query_results_signups.drop_duplicates(subset=['user_name', 'reg_date'], keep='first')

newbiesmask = (qrs['reg_date']>= id_glow_start) 

newbies = qrs.loc[newbiesmask]

len(newbies)

len(editors)
