==> scripts/data_wrangling/8b_edits_at_x_period.ipynb <==
### Edits to the article over time (x period after the contest ends)

# edits_nm_pa_editors_PT18 = wmf.mariadb.run("""
SELECT 
    user_id AS editor, 
    rev_timestamp AS edit_date,
    wiki
FROM user
JOIN revision 
  on rev_user = user_id
JOIN page
  on rev_page = page_id
WHERE user_name IN ({pa_editors_PT18})
  AND rev_timestamp >= "2018-01-01"
  AND rev_timestamp <= "2019-11-12"
  AND rev_minor_edit != 0
""".format(pa_editors_PT18=pa_editors_PT18), "pawiki")

edits_nm_pa_editors_PT18['wiki'] = 'pawiki'
edits_nm_pa_editors_PT18['contest']= 'PT18'

query_vars = dict(
    MWH_SNAPSHOT = "2019-06",
    C_START_DATE = datetime.utcnow().strftime("%Y-%m-%d"),
    M_START_DATE = (datetime.today() - timedelta(days=30)).strftime('%Y-%m'),
    M2_START_DATE = (datetime.today() - timedelta(days=60)).strftime('%Y-%m'),
    Y_START_DATE = (datetime.today() - timedelta(days=365)).strftime('%Y-%m'),
    END_DATE=dt.datetime.now().strftime('%Y-%m'),
)

# six months after contest ends, 2018-11-31

edits_nm_pa_editors_PT18 = wmf.mariadb.run("""
SELECT user_id AS editor, rev_timestamp AS edit_date
FROM user
JOIN revision 
on rev_user = user_id
WHERE user_name IN ({pa_editors_PT18})
  AND rev_timestamp > "20170101000000"
  AND rev_minor_edit != 0
""".format(pa_editors_PT18=pa_editors_PT18), "pawiki")

edits_nm_pa_editors_PT18['wiki'] = 'pawiki'
edits_nm_pa_editors_PT18['contest']= 'PT18'

# df = edits_nm_pa_editors_PT18

# df['edit_date'] = pd.to_datetime(df['edit_date'])

# df.sort_values(by='edit_date').tail()

#save data
df.to_csv("./data/editor_r/editor_retention_PTP_wla_2018.csv", sep=',', encoding = 'utf-8')

# fig, ax = plt.subplots(figsize=(15,7))
df.groupby(['wiki','edit_date']).count()['contest'].unstack().plot(ax=ax)

# df\
    .groupby(['editor','edit_date', 'contest'])\
    .count()\
    .unstack()\
    .plot(kind='bar')

# py.plot([{
    'x': df.index,
    'y':df['contest']
    'name': 'edit_date'
}

for row in df.edit_date])

# df_test = df.copy()

# df_test['edit_count'] = 1

# df_test

# (ggplot(df_test)
    +aes(x='edit_date', y='editor')
 +geom_point()
)

# sns.lmplot(x='edit_count', y='edit_date', data=df_test, col='wiki', hue='contest')

# 

#
