==> scripts/data_wrangling/8c_grantee_data.ipynb <==
import pandas as pd
import wmfdata as wmf
from wmfdata import charting, mariadb, hive
from wmfdata.utils import pct_str, pd_display_all

participants_last_grantees = """
SELECT pp.username, 
pp.incentive_type AS pp_inctv,
gtp.incentive_type AS gtp_inctv,
pp.glow_editor_signup AS pp_editor_signup

FROM florez.glow_tiger_participants gtp
INNER JOIN glow_tiger_pilot_participants pp
on gtp.username = pp.username
WHERE gtp.glow_editor_signup = True
AND pp.incentive_recipient = True
"""

participants_last_participants = """
SELECT pp.username, 
pp.incentive_type AS pp_inctv,
gtp.incentive_type AS gtp_inctv,
pp.glow_editor_signup AS pp_editor_signup
FROM florez.glow_tiger_participants gtp
INNER JOIN glow_tiger_pilot_participants pp
on gtp.username = pp.username
WHERE gtp.glow_editor_signup = True
AND pp.glow_editor_signup = True
"""

from pyspark.sql.types import ArrayType, StringType

from pyspark.context import SparkContext
from pyspark.sql.session import SparkSession

import datetime as dt

import pandas as pd
import numpy as np

non_grantees_last_grantees = """
SELECT count(*)
FROM florez.glow_tiger_pilot_participants pp
    on gtp.username = pp.username
WHERE gtp.glow_editor_signup = True
AND gtp.incentive_recipient = False
AND pp.incentive_recipient = True
AND pp.incentive_type ='laptop'
"""

nglg = spark.sql(non_grantees_last_grantees).toPandas()

contest_articles_by_incentive = '''
SELECT
    count(*)
FROM florez.glow_tiger_articles gta
LEFT JOIN wmf.mediawiki_history mh
    on gta.page_id = mh.page_id
WHERE gta.article_type = "new"
    AND mh.snapshot = '2020-03' 
    AND mh.event_timestamp >= '2019-10-10' 
    AND mh.event_timestamp < '2020-02-10'  
    AND event_entity = "page"
    AND event_type = "create"
GROUP BY database_code
'''
